+++ CHANGES.md
@@ -4,7 +4,7 @@
 - webworker for fhirPathCalculator, which will improve performance for large questionnaires.
 - Inline webworker
 - Made the webworker into a singleton, so that it is only created once and reused for all calculations
-- Change to mocular worker
+- Change to modular worker
 - not use inline webworker, but use a separate worker file
 
 ## 18.1.0-beta01